refactor: consolidate all tests into integration test file
Move inline unit tests from normalize.rs to
tests/event_normalizer_test.rs. All tests use the public API only
(normalize_sse_line) so they don't need module-private access.
Matches the repo convention established by PR #46's cassette tests.
