+++ crates/agentic-core/tests/event_normalizer_test.rs
@@ -1,6 +1,314 @@
 use agentic_core::events::{EventPayload, SSEEventType, normalize_sse_line};
 use serde::Deserialize;
 
+// --- Unit tests (per-event-type parsing) ---
+
+#[test]
+fn test_text_delta() {
+    let line = r#"data: {"type":"response.output_text.delta","delta":"hello","item_id":"msg_1","output_index":0,"content_index":0,"sequence_number":4}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::OutputTextDelta);
+    assert_eq!(frame.sequence_number, Some(4));
+    if let EventPayload::TextDelta {
+        delta,
+        item_id,
+        output_index,
+        content_index,
+    } = &frame.payload
+    {
+        assert_eq!(delta, "hello");
+        assert_eq!(item_id, "msg_1");
+        assert_eq!(*output_index, 0);
+        assert_eq!(*content_index, 0);
+    } else {
+        panic!("expected TextDelta payload");
+    }
+}
+
+#[test]
+fn test_function_call_args_delta() {
+    let line = r#"data: {"type":"response.function_call_arguments.delta","delta":"{\"city\":","call_id":"call_abc","item_id":"fc_1","output_index":0,"sequence_number":7}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::FunctionCallArgumentsDelta);
+    assert_eq!(frame.sequence_number, Some(7));
+    if let EventPayload::FunctionCallArgsDelta {
+        delta,
+        call_id,
+        item_id,
+        output_index,
+    } = &frame.payload
+    {
+        assert_eq!(delta, r#"{"city":"#);
+        assert_eq!(call_id.as_deref(), Some("call_abc"));
+        assert_eq!(item_id, "fc_1");
+        assert_eq!(*output_index, 0);
+    } else {
+        panic!("expected FunctionCallArgsDelta payload");
+    }
+}
+
+#[test]
+fn test_function_call_args_done() {
+    let line = r#"data: {"type":"response.function_call_arguments.done","arguments":"{\"city\":\"SF\"}","call_id":"call_abc","item_id":"fc_1","name":"get_weather","output_index":0,"sequence_number":8}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::FunctionCallArgumentsDone);
+    if let EventPayload::FunctionCallArgsDone {
+        arguments,
+        call_id,
+        name,
+        ..
+    } = &frame.payload
+    {
+        assert_eq!(arguments, r#"{"city":"SF"}"#);
+        assert_eq!(call_id.as_deref(), Some("call_abc"));
+        assert_eq!(name, "get_weather");
+    } else {
+        panic!("expected FunctionCallArgsDone payload");
+    }
+}
+
+#[test]
+fn test_output_item_done() {
+    let line = r#"data: {"type":"response.output_item.done","item":{"id":"msg_1","type":"message","status":"completed","content":[{"type":"output_text","text":"hi"}]},"output_index":0,"sequence_number":9}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::OutputItemDone);
+    if let EventPayload::OutputItemDone {
+        item_id,
+        item_type,
+        item,
+        ..
+    } = &frame.payload
+    {
+        assert_eq!(item_id, "msg_1");
+        assert_eq!(item_type, "message");
+        assert_eq!(item["content"][0]["text"].as_str(), Some("hi"));
+    } else {
+        panic!("expected OutputItemDone payload");
+    }
+}
+
+#[test]
+fn test_vllm_response_done_maps_to_completed() {
+    let line = r#"data: {"type":"response.done","response":{"id":"resp_1","status":"completed","usage":{"total_tokens":10}},"sequence_number":9}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ResponseCompleted);
+    if let EventPayload::Response { id, status, usage } = &frame.payload {
+        assert_eq!(id, "resp_1");
+        assert_eq!(status, "completed");
+        assert!(usage.is_some());
+    } else {
+        panic!("expected Response payload");
+    }
+}
+
+#[test]
+fn test_openai_response_completed() {
+    let line = r#"data: {"type":"response.completed","response":{"id":"resp_2","status":"completed","usage":null},"sequence_number":10}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ResponseCompleted);
+    if let EventPayload::Response { id, usage, .. } = &frame.payload {
+        assert_eq!(id, "resp_2");
+        assert!(usage.is_none());
+    } else {
+        panic!("expected Response payload");
+    }
+}
+
+#[test]
+fn test_done_marker_returns_none() {
+    assert!(normalize_sse_line("data: [DONE]").is_none());
+}
+
+#[test]
+fn test_non_data_lines_return_none() {
+    assert!(normalize_sse_line("event: response.created").is_none());
+    assert!(normalize_sse_line("").is_none());
+    assert!(normalize_sse_line(": comment").is_none());
+    assert!(normalize_sse_line("id: 123").is_none());
+}
+
+#[test]
+fn test_unknown_event_type() {
+    let line = r#"data: {"type":"response.unknown_future_event","foo":"bar"}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::Other);
+    assert!(matches!(frame.payload, EventPayload::Raw(_)));
+}
+
+#[test]
+fn test_malformed_json_returns_none() {
+    assert!(normalize_sse_line("data: {not valid json}").is_none());
+    assert!(normalize_sse_line("data: ").is_none());
+}
+
+#[test]
+fn test_response_created() {
+    let line = r#"data: {"type":"response.created","response":{"id":"resp_abc","status":"in_progress","usage":null},"sequence_number":0}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ResponseCreated);
+    assert_eq!(frame.sequence_number, Some(0));
+    if let EventPayload::Response { id, status, .. } = &frame.payload {
+        assert_eq!(id, "resp_abc");
+        assert_eq!(status, "in_progress");
+    } else {
+        panic!("expected Response payload");
+    }
+}
+
+#[test]
+fn test_output_item_added_message() {
+    let line = r#"data: {"type":"response.output_item.added","item":{"id":"msg_1","type":"message","status":"in_progress","content":[]},"output_index":0,"sequence_number":2}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::OutputItemAdded);
+    if let EventPayload::OutputItemAdded {
+        item_id,
+        item_type,
+        output_index,
+        ..
+    } = &frame.payload
+    {
+        assert_eq!(item_id, "msg_1");
+        assert_eq!(item_type, "message");
+        assert_eq!(*output_index, 0);
+    } else {
+        panic!("expected OutputItemAdded payload");
+    }
+}
+
+#[test]
+fn test_output_item_added_function_call() {
+    let line = r#"data: {"type":"response.output_item.added","item":{"id":"fc_1","type":"function_call","status":"in_progress","name":"get_weather","call_id":"call_1","arguments":""},"output_index":1,"sequence_number":5}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::OutputItemAdded);
+    if let EventPayload::OutputItemAdded {
+        item_id,
+        item_type,
+        output_index,
+        name,
+        call_id,
+    } = &frame.payload
+    {
+        assert_eq!(item_id, "fc_1");
+        assert_eq!(item_type, "function_call");
+        assert_eq!(*output_index, 1);
+        assert_eq!(name.as_deref(), Some("get_weather"));
+        assert_eq!(call_id.as_deref(), Some("call_1"));
+    } else {
+        panic!("expected OutputItemAdded payload");
+    }
+}
+
+#[test]
+fn test_content_part_added_is_raw() {
+    let line = r#"data: {"type":"response.content_part.added","content_index":0,"item_id":"msg_1","output_index":0,"part":{"type":"output_text","text":""},"sequence_number":3}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ContentPartAdded);
+    assert!(matches!(frame.payload, EventPayload::Raw(_)));
+}
+
+#[test]
+fn test_no_sequence_number() {
+    let line =
+        r#"data: {"type":"response.output_text.delta","delta":"x","item_id":"m","output_index":0,"content_index":0}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.sequence_number, None);
+}
+
+#[test]
+fn test_reasoning_delta() {
+    let line = r#"data: {"type":"response.reasoning_summary_text.delta","delta":"Let me think","item_id":"rs_1","sequence_number":3}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ReasoningSummaryTextDelta);
+    if let EventPayload::ReasoningDelta { delta, item_id } = &frame.payload {
+        assert_eq!(delta, "Let me think");
+        assert_eq!(item_id, "rs_1");
+    } else {
+        panic!("expected ReasoningDelta payload");
+    }
+}
+
+#[test]
+fn test_reasoning_done_reads_text_not_delta() {
+    let line = r#"data: {"type":"response.reasoning_summary_text.done","text":"Full reasoning summary here","item_id":"rs_1","sequence_number":5}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ReasoningSummaryTextDone);
+    if let EventPayload::ReasoningDone { text, item_id } = &frame.payload {
+        assert_eq!(text, "Full reasoning summary here");
+        assert_eq!(item_id, "rs_1");
+    } else {
+        panic!("expected ReasoningDone payload");
+    }
+}
+
+#[test]
+fn test_response_failed() {
+    let line = r#"data: {"type":"response.failed","response":{"id":"resp_err","status":"failed","usage":null},"sequence_number":2}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ResponseFailed);
+    if let EventPayload::Response { id, status, .. } = &frame.payload {
+        assert_eq!(id, "resp_err");
+        assert_eq!(status, "failed");
+    } else {
+        panic!("expected Response payload");
+    }
+}
+
+#[test]
+fn test_response_incomplete() {
+    let line = r#"data: {"type":"response.incomplete","response":{"id":"resp_inc","status":"incomplete","usage":{"input_tokens":100,"output_tokens":4096,"total_tokens":4196}},"sequence_number":99}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::ResponseIncomplete);
+    if let EventPayload::Response { status, usage, .. } = &frame.payload {
+        assert_eq!(status, "incomplete");
+        assert!(usage.is_some());
+    } else {
+        panic!("expected Response payload");
+    }
+}
+
+#[test]
+fn test_empty_delta() {
+    let line = r#"data: {"type":"response.output_text.delta","delta":"","item_id":"msg_1","output_index":0,"content_index":0,"sequence_number":4}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::OutputTextDelta);
+    if let EventPayload::TextDelta { delta, .. } = &frame.payload {
+        assert_eq!(delta, "");
+    } else {
+        panic!("expected TextDelta payload");
+    }
+}
+
+#[test]
+fn test_unicode_in_delta() {
+    let line = r#"data: {"type":"response.output_text.delta","delta":"こんにちは 🌍","item_id":"msg_1","output_index":0,"content_index":0,"sequence_number":4}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    if let EventPayload::TextDelta { delta, .. } = &frame.payload {
+        assert_eq!(delta, "こんにちは 🌍");
+    } else {
+        panic!("expected TextDelta payload");
+    }
+}
+
+#[test]
+fn test_file_search_classification() {
+    let line =
+        r#"data: {"type":"response.file_search_call.searching","item_id":"fs_1","output_index":0,"sequence_number":3}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::FileSearchCallSearching);
+    assert!(matches!(frame.payload, EventPayload::Raw(_)));
+}
+
+#[test]
+fn test_web_search_classification() {
+    let line =
+        r#"data: {"type":"response.web_search_call.completed","item_id":"ws_1","output_index":0,"sequence_number":6}"#;
+    let frame = normalize_sse_line(line).unwrap();
+    assert_eq!(frame.event_type, SSEEventType::WebSearchCallCompleted);
+    assert!(matches!(frame.payload, EventPayload::Raw(_)));
+}
+
+// --- Helpers and constants for integration tests ---
+
 const CASSETTE_DIR: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/tests/cassettes/events");
 
 #[derive(Deserialize)]
